refactor: extracted csv path+header validator
